fix not insert yesterday avgPrice

+++ src/main/kotlin/io/olkkani/lfr/dto/MarketApiDTO.kt
@@ -1,11 +1,7 @@
 package io.olkkani.lfr.dto
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import com.fasterxml.jackson.databind.PropertyNamingStrategies
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-import io.olkkani.lfr.entity.jpa.MarketPriceIndex
 import io.olkkani.lfr.entity.mongo.MarketTodayPrice
-import java.time.LocalDate
 
 class MarketRequest(
     val categoryCode: Int,
@@ -19,7 +15,7 @@ data class MarketResponse (
 data class Items (
     @JsonProperty("Id") val id: Int,
     @JsonProperty("CurrentMinPrice") val currentMinPrice: Int,
-    @JsonProperty("YDayAvgPrice") val yDayAvgPrice: Double,
+    @JsonProperty("YDayAvgPrice") val yDayAvgPrice: Float,
 )
 
 fun Items.toMarketTodayPrice() = MarketTodayPrice(
